+++ products/jbrowse-cli/src/commands/text-index.ts
@@ -1,14 +1,11 @@
 import path from 'path'
-import { SingleBar, Presets } from 'cli-progress'
-import readline from 'readline'
 import fs from 'fs'
 import { flags } from '@oclif/command'
 import { Readable } from 'stream'
 import { ixIxxStream } from 'ixixx'
-import { IncomingMessage } from 'http'
-import { http, https, FollowResponse } from 'follow-redirects'
-import { createGunzip } from 'zlib'
 import JBrowseCommand, { Track, Config } from '../base'
+import { indexGff3 } from '../types/gff3Adapter'
+import { indexGtf } from '../types/gtfAdapter'
 
 export default class TextIndex extends JBrowseCommand {
   static description = 'Make a text-indexing file for any given track(s).'
@@ -189,192 +186,11 @@ export default class TextIndex extends JBrowseCommand {
       } = config
 
       if (uri.endsWith('.gtf')) {
-        yield* this.indexGtf(config, attributes, outLocation, quiet)
+        yield* indexGtf(config, attributes, outLocation, quiet)
       } else if (type === 'Gff3TabixAdapter') {
-        yield* this.indexGff3(config, attributes, outLocation, quiet)
+        yield* indexGff3(config, attributes, outLocation, quiet)
       }
     }
-  }
-
-  async *indexGtf(
-    config: Track,
-    attributes: string[],
-    outLocation: string,
-    quiet: boolean,
-  ) {
-    const {
-      adapter: {
-        gffGzLocation: { uri },
-      },
-      trackId,
-    } = config
-    // progress bar code was aided by blog post at
-    // https://webomnizz.com/download-a-file-with-progressbar-using-node-js/
-    const progressBar = new SingleBar(
-      {
-        format: '{bar} ' + trackId + ' {percentage}% | ETA: {eta}s',
-      },
-      Presets.shades_classic,
-    )
-
-    let fileDataStream
-    let totalBytes = 0
-    let receivedBytes = 0
-    if (this.isURL(uri)) {
-      fileDataStream = await this.createRemoteStream(uri)
-      totalBytes = +(fileDataStream.headers['content-length'] || 0)
-    } else {
-      const filename = path.join(outLocation, uri)
-      totalBytes = fs.statSync(filename).size
-      fileDataStream = fs.createReadStream(filename)
-    }
-
-    if (!quiet) {
-      progressBar.start(totalBytes, 0)
-    }
-
-    fileDataStream.on('data', chunk => {
-      receivedBytes += chunk.length
-      progressBar.update(receivedBytes)
-    })
-
-    const gzStream = uri.endsWith('.gz')
-      ? fileDataStream.pipe(createGunzip())
-      : fileDataStream
-
-    const rl = readline.createInterface({
-      input: gzStream,
-    })
-
-    const regex = /(["])/g
-
-    for await (const line of rl) {
-      if (line.startsWith('#')) {
-        continue
-      } else if (line.startsWith('>')) {
-        break
-      }
-
-      const [seq_name, , , start, end, , , , col9] = line.split('\t')
-      const locStr = `${seq_name}:${start}..${end}`
-
-      const col9attrs = col9.split('; ')
-      const attrs = attributes.map(attr =>
-        col9attrs
-          .find(f => f.startsWith(attr))
-          ?.split(' ')[1]
-          .trim()
-          .replace(regex, ''),
-      )
-      const record = JSON.stringify([locStr, trackId])
-
-      const buff = Buffer.from(record).toString('base64')
-      yield `${buff} ${[...new Set(attrs)].join(' ')}\n`
-    }
-    progressBar.stop()
-  }
-
-  async *indexGff3(
-    config: Track,
-    attributes: string[],
-    outLocation: string,
-    quiet: boolean,
-  ) {
-    const {
-      adapter: {
-        gffGzLocation: { uri },
-      },
-      trackId,
-    } = config
-
-    // progress bar code was aided by blog post at
-    // https://webomnizz.com/download-a-file-with-progressbar-using-node-js/
-    const progressBar = new SingleBar(
-      {
-        format: '{bar} ' + trackId + ' {percentage}% | ETA: {eta}s',
-      },
-      Presets.shades_classic,
-    )
-
-    let fileDataStream
-    let totalBytes = 0
-    let receivedBytes = 0
-    if (this.isURL(uri)) {
-      fileDataStream = await this.createRemoteStream(uri)
-      totalBytes = +(fileDataStream.headers['content-length'] || 0)
-    } else {
-      const filename = path.join(outLocation, uri)
-      totalBytes = fs.statSync(filename).size
-      fileDataStream = fs.createReadStream(filename)
-    }
-
-    if (!quiet) {
-      progressBar.start(totalBytes, 0)
-    }
-
-    fileDataStream.on('data', chunk => {
-      receivedBytes += chunk.length
-      progressBar.update(receivedBytes)
-    })
-
-    const gzStream = uri.endsWith('.gz')
-      ? fileDataStream.pipe(createGunzip())
-      : fileDataStream
-
-    const rl = readline.createInterface({
-      input: gzStream,
-    })
-
-    for await (const line of rl) {
-      if (line.startsWith('#')) {
-        continue
-      } else if (line.startsWith('>')) {
-        break
-      }
-
-      const [seq_id, , , start, end, , , , col9] = line.split('\t')
-      const locStr = `${seq_id}:${start}..${end}`
-
-      const col9attrs = col9.split(';')
-      const attrs = attributes.map(attr =>
-        col9attrs
-          .find(f => f.startsWith(attr))
-          ?.split('=')[1]
-          .trim(),
-      )
-      const record = JSON.stringify([locStr, trackId])
-
-      const buff = Buffer.from(record).toString('base64')
-      yield `${buff} ${[...new Set(attrs)].join(' ')}\n`
-    }
-
-    progressBar.stop()
-  }
-
-  // Method for handing off the parsing of a gff3 file URL.
-  // Calls the proper parser depending on if it is gzipped or not.
-  // Returns a @gmod/gff stream.
-  async createRemoteStream(urlIn: string) {
-    const newUrl = new URL(urlIn)
-    const fetcher = newUrl.protocol === 'https:' ? https : http
-
-    return new Promise<IncomingMessage & FollowResponse>((resolve, reject) =>
-      fetcher.get(urlIn, resolve).on('error', reject),
-    )
-  }
-
-  // Checks if the passed in string is a valid URL.
-  // Returns a boolean.
-  isURL(FileName: string) {
-    let url
-
-    try {
-      url = new URL(FileName)
-    } catch (_) {
-      return false
-    }
-
-    return url.protocol === 'http:' || url.protocol === 'https:'
   }
 
   // Given a readStream of data, indexes the stream into .ix and .ixx files
