+++ products/jbrowse-cli/src/types/common.ts
@@ -1,0 +1,28 @@
+import { IncomingMessage } from 'http'
+import { http, https, FollowResponse } from 'follow-redirects'
+
+// Method for handing off the parsing of a gff3 file URL.
+// Calls the proper parser depending on if it is gzipped or not.
+// Returns a @gmod/gff stream.
+export async function createRemoteStream(urlIn: string) {
+  const newUrl = new URL(urlIn)
+  const fetcher = newUrl.protocol === 'https:' ? https : http
+
+  return new Promise<IncomingMessage & FollowResponse>((resolve, reject) =>
+    fetcher.get(urlIn, resolve).on('error', reject),
+  )
+}
+
+// Checks if the passed in string is a valid URL.
+// Returns a boolean.
+export function isURL(FileName: string) {
+  let url
+
+  try {
+    url = new URL(FileName)
+  } catch (_) {
+    return false
+  }
+
+  return url.protocol === 'http:' || url.protocol === 'https:'
+}
