+++ products/jbrowse-cli/test/data/config.json
@@ -51,6 +51,25 @@
   "tracks": [
     {
       "type": "FeatureTrack",
+      "trackId": "gtftabix_genes",
+      "assemblyNames": ["volvox"],
+      "name": "Gtf Tabix Genes",
+      "category": ["Miscellaneous"],
+      "adapter": {
+        "type": "Gff3TabixAdapter",
+        "gffGzLocation": {
+          "uri": "Felis_catus.Felis_catus_9.0.104.chr.gtf"
+        },
+        "index": {
+          "location": {
+            "uri": "Felis_catus.Felis_catus_9.0.104.chr.gtf.tbi"
+          }
+        }
+      },
+      "textSearchIndexingAttributes": ["gene_name", "gene_id", "feature"]
+    },
+    {
+      "type": "FeatureTrack",
       "trackId": "au9_scaffold",
       "assemblyNames": ["volvox"],
       "name": "AU9 Scaffold",
